[Reactify] Data retrieval a hook
