+++ aquawatch_mobile_app/src/hooks/index.js
@@ -1,1 +1,2 @@
 export { default as useLocationMetaProvider } from './useLocationMetaProvider';
+export { default as useGetWaterData } from './useGetWaterData';
