+++ aquawatch_mobile_app/src/hooks/useLocationMetaProvider.js
@@ -3,7 +3,8 @@ import stationIds from './jsons/stationIds.json';
 import usgsParameterMappings from './jsons/usgsParameterMapping.json';
 import locationOptions from './jsons/locationOptions.json';
 
-export default function useLocationMetaProvider() {    
+export default function useLocationMetaProvider() {   
+
     return {
         parameterInfo,
         stationIds,
